+++ docs/release-notes.md
@@ -15,6 +15,15 @@ hide:
 - Fix crash with `auto_now_add` as primary key.
 - Fix "$defs" not matching def names. This fixes problems with enums which appear multiple times and reduces requests.
 - Fix using ForeignKeys in embedded models/CompositeFields when the embedded model has no registry set (e.g. abstract).
+- Improve admin UI.
+- Add some extra filters and option to make Primarykeys read-only in marshall.
+
+### Changed
+
+- Remove by default read-only fields from admin create/update.
+- Remove by default autoincrement fields from admin create.
+- Make PrimaryKeys read-only for update (they will be removed because they are read-only until we have a better solution).
+- `FileStruct` name have a minimal length of 1.
 
 ## 0.32.4